test(ingester): split write / query tests
Split the write & query integration tests into their own modules for
clarity.
